+++ CHANGELOG.md
@@ -18,12 +18,23 @@ Categories, defined in [changemap.json](.github/clq/changemap.json):
   - `Fixed` for any bugfixes.
   - `Security` in case of vulnerabilities.
 
+## [1.2.0] - 2025-01-19
+
+### Added
+
+- Export the name and version of the chart as `chart_name`, `chart_version` if this action published it.
+- Registries are now passed to the action as input
+
+#### Fixed
+
+- Fix build badge to refer to the new build filename
+
 ## [1.1.0] - 2025-01-09
 
 ### Added
 
-- publish single helm chart from `build/helm/charts/' if on the main branch
-- publish single docker image from `build/jib-image.tar' if on the main branch
+- publish single helm chart from `build/helm/charts/` if on the main branch
+- publish single docker image from `build/jib-image.tar` if on the main branch
 
 ### Fixed